+++ packages/tkeel-console-components/src/components/Toast/ToastContainer.tsx
@@ -1,19 +1,17 @@
-import { useTheme } from '@chakra-ui/react';
 import { merge } from 'lodash';
 import {
   ToastContainer as ToastifyToastContainer,
   ToastContainerProps,
 } from 'react-toastify';
 
-import { useStatusInfos } from '@tkeel/console-hooks';
-import { Theme } from '@tkeel/console-themes';
+import { useColors, useStatusInfos } from '@tkeel/console-hooks';
 
 import { DEFAULT_TOAST_CONTAINER_PROPS } from './constants';
 import { StyledWrapper } from './styled';
 
 export default function ToastContainer(props: ToastContainerProps) {
   const properties = merge({}, DEFAULT_TOAST_CONTAINER_PROPS, props);
-  const { colors }: Theme = useTheme();
+  const colors = useColors();
   const statusInfos = useStatusInfos();
 
   return (
