+++ packages/tkeel-console-constants/src/constants/portals/index.ts
@@ -1,0 +1,15 @@
+export const enum PortalNames {
+  ADMIN = 'admin',
+  TENANT = 'tenant',
+}
+
+export const PORTAL_INFOS = {
+  [PortalNames.ADMIN]: {
+    name: PortalNames.ADMIN,
+  },
+  [PortalNames.TENANT]: {
+    name: PortalNames.TENANT,
+  },
+};
+
+export const DEFAULT_PORTAL_NAME = PortalNames.TENANT;
