+++ packages/tkeel-console-hooks/src/hooks/useColors/index.ts
@@ -3,7 +3,7 @@ import { useTheme } from '@chakra-ui/react';
 import type { Theme } from '@tkeel/console-themes';
 
 export default function useColors() {
-  const theme: Theme = useTheme();
+  const theme = useTheme<Theme>();
 
   return theme.colors;
 }
