+++ packages/tkeel-console-hooks/src/hooks/useStatusInfos/index.tsx
@@ -1,9 +1,9 @@
 import { useTheme } from '@chakra-ui/react';
 
-import { Theme } from '@tkeel/console-themes';
+import type { Theme } from '@tkeel/console-themes';
 import { getStatusInfos } from '@tkeel/console-utils';
 
 export default function useStatusInfos() {
-  const { colors }: Theme = useTheme();
+  const { colors } = useTheme<Theme>();
   return getStatusInfos({ colors });
 }
