+++ packages/tkeel-console-plugin-admin-service-monitoring/src/hooks/usePluginStatusInfo/index.ts
@@ -1,6 +1,4 @@
-import { useTheme } from '@chakra-ui/react';
-
-import { Theme } from '@tkeel/console-themes';
+import { useColors } from '@tkeel/console-hooks';
 
 import { getPluginStatusInfo } from '@/tkeel-console-plugin-admin-service-monitoring/constants/plugins';
 import type { PluginStatus } from '@/tkeel-console-plugin-admin-service-monitoring/types';
@@ -10,6 +8,6 @@ export default function usePluginStatusInfo({
 }: {
   status: PluginStatus;
 }) {
-  const { colors }: Theme = useTheme();
+  const colors = useColors();
   return getPluginStatusInfo({ status, colors });
 }
