+++ packages/tkeel-console-plugin-admin-service-monitoring/src/hooks/usePodStatusInfo/index.ts
@@ -1,11 +1,9 @@
-import { useTheme } from '@chakra-ui/react';
-
-import { Theme } from '@tkeel/console-themes';
+import { useColors } from '@tkeel/console-hooks';
 
 import { getPodStatusInfo } from '@/tkeel-console-plugin-admin-service-monitoring/constants/pods';
 import type { PodStatus } from '@/tkeel-console-plugin-admin-service-monitoring/types';
 
 export default function usePodStatusInfo({ status }: { status: PodStatus }) {
-  const { colors }: Theme = useTheme();
+  const colors = useColors();
   return getPodStatusInfo({ status, colors });
 }
