+++ packages/tkeel-console-portal-tenant/config/default.js
@@ -4,6 +4,7 @@ module.exports = {
   portalName: 'tenant',
   publicPath: '/',
   client: {
+    brand: 'tkeel', // qingcloud
     themeName: 'qingcloud-light',
   },
   backend: {
