+++ packages/tkeel-console-utils/src/utils/status/types.ts
@@ -1,7 +1,7 @@
 import { FunctionComponent } from 'react';
 
 import { TwoToneIconProps } from '@tkeel/console-icons';
-import { Theme } from '@tkeel/console-themes';
+import type { Theme } from '@tkeel/console-themes';
 
 type StatusKeys = 'default' | 'info' | 'success' | 'warning' | 'error';
